+++ lib/std_library.dart
@@ -8,14 +8,15 @@ var environment = {
   'modulo': modulo,
   'pow': pow,
   'max': max,
-  'log': log,
+  'print': log,
   "pi": {
     "kind": "final",
     "value": () => {
           "type": "double",
           "value": pi,
         }
-  }
+  },
+  "modules": []
 };
 
 all(Function(dynamic, dynamic) fn) => (List list) => list.reduce(fn);
@@ -75,7 +76,7 @@ log(Map params) {
   var left = params['left'];
   var right = params['right'];
 
-  if (right == null || right.isEmpty) {
+  if (right == null || right.isEmpty || right[0] == null) {
     print(left.toString());
     return;
   }